requirements.txt
gymnasium
numpy
matplotlib
torch
ruamel.yaml
mujoco
mujoco-py
